# Conflicts:
#	src/main/java/com/jll/pay/order/DepositOrderServiceImpl.java

+++ src/main/java/com/jll/game/playtypefacade/EleIn5QwZwPlayTypeFacadeImpl.java
@@ -97,7 +97,7 @@ public class EleIn5QwZwPlayTypeFacadeImpl extends DefaultPlayTypeFacadeImpl {
 		}
 		
 		betAmount = MathUtil.multiply(betTotal, times, Float.class);
-		betAmount = MathUtil.multiply(betAmount, monUnit, Float.class);
+		betAmount = MathUtil.multiply(betAmount, monUnit.floatValue(), Float.class);
 		maxWinAmount = MathUtil.multiply(betAmount, singleBettingPrize.floatValue(), Float.class);
 		
 		ret.put("playType", playType);
